LayoutTests/fast/scrolling/mac/keyboard-scrolling-pagedown-pageup-expected.txt

PASS scrollOffsetAfterPageDown > scrollOffsetAfterPageUp is true
PASS successfullyParsed is true

TEST COMPLETE
